refactor model.py into components
