+++ torchtitan/experiments/deepseek_v3/models/model.py
@@ -12,1396 +12,68 @@
 # It has been modified from its original forms to accommodate naming convention
 # and usage patterns of the TorchTitan project.
 
-# Copyright 2023 DeepSeek-AI and The HuggingFace Inc. team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
 """ PyTorch DeepSeek model."""
 
+# Define all exported symbols
+__all__ = [
+    # Normalization
+    "RMSNorm",
+    # Rotary Embeddings
+    "RotaryEmbedding",
+    "LinearScalingRotaryEmbedding",
+    "DynamicNTKScalingRotaryEmbedding",
+    "YarnRotaryEmbedding",
+    "rotate_half",
+    "apply_rotary_pos_emb",
+    "yarn_find_correction_dim",
+    "yarn_find_correction_range",
+    "yarn_get_mscale",
+    "yarn_linear_ramp_mask",
+    # MLP and MoE
+    "MLP",
+    "MoE",
+    "MoEGate",
+    "get_group",
+    # Attention
+    "Attention",
+    "_prepare_4d_causal_attention_mask",
+    # Model Architecture
+    "DecoderLayer",
+    "DeepseekModel",
+    "DeepseekForCausalLM",
+    # Configuration
+    "ModelArgs",
+    "deepseek_config_registry",
+    "deepseek_v2_lite_config",
+    # Backward compatibility
+    "model",
+    "model_for_causal_lm",
+]
 
-import math
+# Re-export all components from their respective modules
 
-# from re import X
-from typing import Optional, Tuple
+from .attention import Attention
+from .attn_mask_utils import _prepare_4d_causal_attention_mask
+from .mlp import MLP
+from .model_architecture import DecoderLayer, DeepseekForCausalLM, DeepseekModel
+from .model_config import deepseek_config_registry, deepseek_v2_lite_config, ModelArgs
+from .moe import get_group, MoE, MoEGate
 
-import torch
-import torch.distributed as dist
-
-import torch.distributed._symmetric_memory as symm_mem
-import torch.nn.functional as F
-import torch.utils.checkpoint
-from symm_mem_recipes import OnDeviceAllToAllV
-from torch import nn
-from torch.distributed._functional_collectives import all_to_all_single_autograd
-
-from torchtitan.experiments.kernels.group_gemms.group_gemms import (
-    DSGroupGEMM,
-    TorchAOBF16GroupGEMM,
-    TorchBF16GroupGEMM,
-    TorchFP8GroupGEMM,
-    TritonCGBF16GroupGEMM,
+# Re-export all components from their respective modules
+from .normalization import RMSNorm
+from .rope_embeddings import (
+    apply_rotary_pos_emb,
+    DynamicNTKScalingRotaryEmbedding,
+    LinearScalingRotaryEmbedding,
+    RotaryEmbedding,
+    rotate_half,
+    yarn_find_correction_dim,
+    yarn_find_correction_range,
+    yarn_get_mscale,
+    yarn_linear_ramp_mask,
+    YarnRotaryEmbedding,
 )
 
-from torchtitan.experiments.kernels.moe.indices import generate_permute_indices
-from torchtitan.experiments.kernels.triton_mg_group_gemm.torchao_pr import ALIGN_SIZE_M
-
-from .attn_mask_utils import _prepare_4d_causal_attention_mask
-
-from .model_config import ModelArgs
-
-
-# Get model parallel subgroup by name:
-# e.g. "pp", "ep", None
-def get_group(dim_name: Optional[str] = None) -> dist.ProcessGroup:
-    glob = torch.distributed.device_mesh._mesh_resources.get_current_mesh()
-    return glob.get_group(dim_name)
-
-
-class RMSNorm(nn.Module):
-    def __init__(self, hidden_size, eps=1e-6):
-        super().__init__()
-        self.weight = nn.Parameter(torch.ones(hidden_size))
-        self.variance_epsilon = eps
-
-    def forward(self, hidden_states):
-        input_dtype = hidden_states.dtype
-        hidden_states = hidden_states.to(torch.float32)
-        variance = hidden_states.pow(2).mean(-1, keepdim=True)
-        hidden_states = hidden_states * torch.rsqrt(variance + self.variance_epsilon)
-        return self.weight * hidden_states.to(input_dtype)
-
-
-class RotaryEmbedding(nn.Module):
-    def __init__(self, dim, max_position_embeddings=2048, base=10000, device=None):
-        super().__init__()
-
-        self.dim = dim
-        self.max_position_embeddings = max_position_embeddings
-        self.base = base
-        inv_freq = 1.0 / (
-            self.base ** (torch.arange(0, self.dim, 2).float().to(device) / self.dim)
-        )
-        self.register_buffer("inv_freq", inv_freq, persistent=False)
-
-        # Build here to make `torch.jit.trace` work.
-        self._set_cos_sin_cache(
-            seq_len=max_position_embeddings,
-            device=self.inv_freq.device,
-            dtype=torch.get_default_dtype(),
-        )
-
-    def _set_cos_sin_cache(self, seq_len, device, dtype):
-        self.max_seq_len_cached = seq_len
-        t = torch.arange(
-            self.max_seq_len_cached, device=device, dtype=self.inv_freq.dtype
-        )
-
-        freqs = torch.outer(t, self.inv_freq.to(t.device))
-        # Different from paper, but it uses a different permutation in order to obtain the same calculation
-        emb = torch.cat((freqs, freqs), dim=-1)
-        self.register_buffer("cos_cached", emb.cos().to(dtype), persistent=False)
-        self.register_buffer("sin_cached", emb.sin().to(dtype), persistent=False)
-
-    def forward(self, x, seq_len=None):
-        # x: [bs, num_attention_heads, seq_len, head_size]
-        if self.max_seq_len_cached is None or seq_len > self.max_seq_len_cached:
-            self._set_cos_sin_cache(seq_len=seq_len, device=x.device, dtype=x.dtype)
-
-        return (
-            self.cos_cached[:seq_len].to(dtype=x.dtype),
-            self.sin_cached[:seq_len].to(dtype=x.dtype),
-        )
-
-
-class LinearScalingRotaryEmbedding(RotaryEmbedding):
-    """RotaryEmbedding extended with linear scaling. Credits to the Reddit user /u/kaiokendev"""
-
-    def __init__(
-        self,
-        dim,
-        max_position_embeddings=2048,
-        base=10000,
-        device=None,
-        scaling_factor=1.0,
-    ):
-        self.scaling_factor = scaling_factor
-        super().__init__(dim, max_position_embeddings, base, device)
-
-    def _set_cos_sin_cache(self, seq_len, device, dtype):
-        self.max_seq_len_cached = seq_len
-        t = torch.arange(
-            self.max_seq_len_cached, device=device, dtype=self.inv_freq.dtype
-        )
-        t = t / self.scaling_factor
-
-        freqs = torch.outer(t, self.inv_freq)
-        # Different from paper, but it uses a different permutation in order to obtain the same calculation
-        emb = torch.cat((freqs, freqs), dim=-1)
-        self.register_buffer("cos_cached", emb.cos().to(dtype), persistent=False)
-        self.register_buffer("sin_cached", emb.sin().to(dtype), persistent=False)
-
-
-# Copied from transformers.models.llama.modeling_llama.LlamaDynamicNTKScalingRotaryEmbedding with Llama->Deepseek
-class DynamicNTKScalingRotaryEmbedding(RotaryEmbedding):
-    """RotaryEmbedding extended with Dynamic NTK scaling. Credits to the Reddit users /u/bloc97 and /u/emozilla"""
-
-    def __init__(
-        self,
-        dim,
-        max_position_embeddings=2048,
-        base=10000,
-        device=None,
-        scaling_factor=1.0,
-    ):
-        self.scaling_factor = scaling_factor
-        super().__init__(dim, max_position_embeddings, base, device)
-
-    def _set_cos_sin_cache(self, seq_len, device, dtype):
-        self.max_seq_len_cached = seq_len
-
-        if seq_len > self.max_position_embeddings:
-            base = self.base * (
-                (self.scaling_factor * seq_len / self.max_position_embeddings)
-                - (self.scaling_factor - 1)
-            ) ** (self.dim / (self.dim - 2))
-            inv_freq = 1.0 / (
-                base ** (torch.arange(0, self.dim, 2).float().to(device) / self.dim)
-            )
-            self.register_buffer("inv_freq", inv_freq, persistent=False)
-
-        t = torch.arange(
-            self.max_seq_len_cached, device=device, dtype=self.inv_freq.dtype
-        )
-
-        freqs = torch.outer(t, self.inv_freq)
-        # Different from paper, but it uses a different permutation in order to obtain the same calculation
-        emb = torch.cat((freqs, freqs), dim=-1)
-        self.register_buffer("cos_cached", emb.cos().to(dtype), persistent=False)
-        self.register_buffer("sin_cached", emb.sin().to(dtype), persistent=False)
-
-
-# Inverse dim formula to find dim based on number of rotations
-def yarn_find_correction_dim(
-    num_rotations, dim, base=10000, max_position_embeddings=2048
-):
-    return (dim * math.log(max_position_embeddings / (num_rotations * 2 * math.pi))) / (
-        2 * math.log(base)
-    )
-
-
-# Find dim range bounds based on rotations
-def yarn_find_correction_range(
-    low_rot, high_rot, dim, base=10000, max_position_embeddings=2048
-):
-    low = math.floor(
-        yarn_find_correction_dim(low_rot, dim, base, max_position_embeddings)
-    )
-    high = math.ceil(
-        yarn_find_correction_dim(high_rot, dim, base, max_position_embeddings)
-    )
-    return max(low, 0), min(high, dim - 1)  # Clamp values just in case
-
-
-def yarn_get_mscale(scale=1, mscale=1):
-    if scale <= 1:
-        return 1.0
-    return 0.1 * mscale * math.log(scale) + 1.0
-
-
-def yarn_linear_ramp_mask(min, max, dim):
-    if min == max:
-        max += 0.001  # Prevent singularity
-
-    linear_func = (torch.arange(dim, dtype=torch.float32) - min) / (max - min)
-    ramp_func = torch.clamp(linear_func, 0, 1)
-    return ramp_func
-
-
-class YarnRotaryEmbedding(RotaryEmbedding):
-    def __init__(
-        self,
-        dim,
-        max_position_embeddings=2048,
-        base=10000,
-        device=None,
-        scaling_factor=1.0,
-        original_max_position_embeddings=4096,
-        beta_fast=32,
-        beta_slow=1,
-        mscale=1,
-        mscale_all_dim=0,
-    ):
-        self.scaling_factor = scaling_factor
-        self.original_max_position_embeddings = original_max_position_embeddings
-        self.beta_fast = beta_fast
-        self.beta_slow = beta_slow
-        self.mscale = mscale
-        self.mscale_all_dim = mscale_all_dim
-        super().__init__(dim, max_position_embeddings, base, device)
-
-    def _set_cos_sin_cache(self, seq_len, device, dtype):
-        self.max_seq_len_cached = seq_len
-        dim = self.dim
-
-        freq_extra = 1.0 / (
-            self.base
-            ** (torch.arange(0, dim, 2, dtype=torch.float32, device=device) / dim)
-        )
-        freq_inter = 1.0 / (
-            self.scaling_factor
-            * self.base
-            ** (torch.arange(0, dim, 2, dtype=torch.float32, device=device) / dim)
-        )
-
-        low, high = yarn_find_correction_range(
-            self.beta_fast,
-            self.beta_slow,
-            dim,
-            self.base,
-            self.original_max_position_embeddings,
-        )
-        inv_freq_mask = 1.0 - yarn_linear_ramp_mask(low, high, dim // 2).to(
-            device=device, dtype=torch.float32
-        )
-        inv_freq = freq_inter * (1 - inv_freq_mask) + freq_extra * inv_freq_mask
-        self.register_buffer("inv_freq", inv_freq, persistent=False)
-
-        t = torch.arange(seq_len, device=device, dtype=torch.float32)
-
-        freqs = torch.outer(t, inv_freq)
-
-        _mscale = float(
-            yarn_get_mscale(self.scaling_factor, self.mscale)
-            / yarn_get_mscale(self.scaling_factor, self.mscale_all_dim)
-        )
-
-        emb = torch.cat((freqs, freqs), dim=-1)
-        self.register_buffer(
-            "cos_cached", (emb.cos() * _mscale).to(dtype), persistent=False
-        )
-        self.register_buffer(
-            "sin_cached", (emb.sin() * _mscale).to(dtype), persistent=False
-        )
-
-
-# Copied from transformers.models.llama.modeling_llama.rotate_half
-def rotate_half(x):
-    """Rotates half the hidden dims of the input."""
-    x1 = x[..., : x.shape[-1] // 2]
-    x2 = x[..., x.shape[-1] // 2 :]
-    return torch.cat((-x2, x1), dim=-1)
-
-
-# Copied from transformers.models.llama.modeling_llama.apply_rotary_pos_emb
-def apply_rotary_pos_emb(q, k, cos, sin, position_ids, unsqueeze_dim=1):
-    """Applies Rotary Position Embedding to the query and key tensors.
-
-    Args:
-        q (`torch.Tensor`): The query tensor.
-        k (`torch.Tensor`): The key tensor.
-        cos (`torch.Tensor`): The cosine part of the rotary embedding.
-        sin (`torch.Tensor`): The sine part of the rotary embedding.
-        position_ids (`torch.Tensor`):
-            The position indices of the tokens corresponding to the query and key tensors. For example, this can be
-            used to pass offsetted position ids when working with a KV-cache.
-        unsqueeze_dim (`int`, *optional*, defaults to 1):
-            The 'unsqueeze_dim' argument specifies the dimension along which to unsqueeze cos[position_ids] and
-            sin[position_ids] so that they can be properly broadcasted to the dimensions of q and k. For example, note
-            that cos[position_ids] and sin[position_ids] have the shape [batch_size, seq_len, head_dim]. Then, if q and
-            k have the shape [batch_size, heads, seq_len, head_dim], then setting unsqueeze_dim=1 makes
-            cos[position_ids] and sin[position_ids] broadcastable to the shapes of q and k. Similarly, if q and k have
-            the shape [batch_size, seq_len, heads, head_dim], then set unsqueeze_dim=2.
-    Returns:
-        `tuple(torch.Tensor)` comprising of the query and key tensors rotated using the Rotary Position Embedding.
-    """
-    cos = cos[position_ids].unsqueeze(unsqueeze_dim)
-    sin = sin[position_ids].unsqueeze(unsqueeze_dim)
-
-    b, h, s, d = q.shape
-    q = q.view(b, h, s, d // 2, 2).transpose(4, 3).reshape(b, h, s, d)
-
-    b, h, s, d = k.shape
-    k = k.view(b, h, s, d // 2, 2).transpose(4, 3).reshape(b, h, s, d)
-
-    q_embed = (q * cos) + (rotate_half(q) * sin)
-    k_embed = (k * cos) + (rotate_half(k) * sin)
-    return q_embed, k_embed
-
-
-class MLP(nn.Module):
-    act_fn = nn.SiLU()
-
-    def __init__(self, config, hidden_size=None, intermediate_size=None):
-        super().__init__()
-        self.config = config
-        self.hidden_size = config.hidden_size if hidden_size is None else hidden_size
-        self.intermediate_size = (
-            config.intermediate_size if intermediate_size is None else intermediate_size
-        )
-
-        self.gate_proj = nn.Linear(self.hidden_size, self.intermediate_size, bias=False)
-        self.up_proj = nn.Linear(self.hidden_size, self.intermediate_size, bias=False)
-        self.down_proj = nn.Linear(self.intermediate_size, self.hidden_size, bias=False)
-
-    def forward(self, x):
-        down_proj = self.down_proj(self.act_fn(self.gate_proj(x)) * self.up_proj(x))
-        return down_proj
-
-
-class MoEGate(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        self.config = config
-        self.top_k = config.num_experts_per_tok
-        self.n_routed_experts = config.n_routed_experts
-        self.routed_scaling_factor = config.routed_scaling_factor
-        self.scoring_func = config.scoring_func
-        self.seq_aux = config.seq_aux
-        self.topk_method = config.topk_method
-        self.n_group = config.n_group
-        self.topk_group = config.topk_group
-
-        # topk selection algorithm
-        self.norm_topk_prob = config.norm_topk_prob
-        self.gating_dim = config.hidden_size
-        self.weight = nn.Parameter(
-            torch.empty((self.n_routed_experts, self.gating_dim))
-        )
-        if self.topk_method == "noaux_tc":
-            self.e_score_correction_bias = nn.Parameter(
-                # Changed from torch.empty to torch.rand to avoid non-even
-                # distribution for runs without actual weigths
-                torch.rand((self.n_routed_experts))
-            )
-        self.reset_parameters()
-
-    def reset_parameters(self) -> None:
-        import torch.nn.init as init
-
-        init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-
-    def forward(self, hidden_states):
-        bsz, seq_len, h = hidden_states.shape
-        # compute gating score
-        hidden_states = hidden_states.view(-1, h)
-        logits = F.linear(
-            hidden_states.type(torch.float32), self.weight.type(torch.float32), None
-        )
-        if self.scoring_func == "sigmoid":
-            scores = logits.sigmoid()
-        elif self.scoring_func == "softmax":
-            scores = logits.softmax(dim=-1, dtype=torch.float32)
-        else:
-            raise NotImplementedError(
-                f"insupportable scoring function for MoE gating: {self.scoring_func}"
-            )
-
-        # select top-k experts
-        if self.topk_method == "noaux_tc":
-            scores_for_choice = scores.view(
-                bsz * seq_len, -1
-            ) + self.e_score_correction_bias.unsqueeze(0)
-            group_scores = (
-                scores_for_choice.view(bsz * seq_len, self.n_group, -1)
-                .topk(2, dim=-1)[0]
-                .sum(dim=-1)
-            )  # [n, n_group]
-            group_idx = torch.topk(
-                group_scores, k=self.topk_group, dim=-1, sorted=False
-            )[
-                1
-            ]  # [n, top_k_group]
-            group_mask = torch.zeros_like(group_scores)  # [n, n_group]
-            group_mask.scatter_(1, group_idx, 1)  # [n, n_group]
-            score_mask = (
-                group_mask.unsqueeze(-1)
-                .expand(
-                    bsz * seq_len, self.n_group, self.n_routed_experts // self.n_group
-                )
-                .reshape(bsz * seq_len, -1)
-            )  # [n, e]
-            tmp_scores = scores_for_choice.masked_fill(
-                ~score_mask.bool(), 0.0
-            )  # [n, e]
-            _, topk_idx = torch.topk(tmp_scores, k=self.top_k, dim=-1, sorted=False)
-            topk_weight = scores.gather(1, topk_idx)
-        elif self.topk_method == "greedy":
-            topk_weight, topk_idx = torch.topk(
-                scores, k=self.top_k, dim=-1, sorted=False
-            )
-        else:
-            raise NotImplementedError(
-                f"insupportable TopK function for MoE gating: {self.topk_method}"
-            )
-
-        # norm gate to sum 1
-        if self.top_k > 1 and self.norm_topk_prob:
-            denominator = topk_weight.sum(dim=-1, keepdim=True) + 1e-20
-            topk_weight = topk_weight / denominator
-        topk_weight = (
-            topk_weight * self.routed_scaling_factor
-        )  # must multiply the scaling factor
-
-        return topk_idx, topk_weight
-
-
-class MoE(nn.Module):
-    """
-    A mixed expert module containing shared experts.
-    """
-
-    # Class attributes:
-    # Two shuffle method supported:
-    # 1. "torch_all_to_all"
-    # 2. "symm_mem" (see `setup_symm_mem` below)
-    shuffle_method = "torch_all_to_all"
-
-    # Symmetric memory buffers shared by all MoE instances across layers
-    token_send_buf: Optional[torch.Tensor] = None
-    token_gather_buf: Optional[torch.Tensor] = None
-
-    # Group GEMM strategies
-    group_gemm_strategies = None
-    # which group gemm to use?
-    group_mm = "torch"  # fp8 options = ["torchfp8", "dsgemm"] bf16 = ["torch", , "torchao", "tritoncg"]
-
-    def __init__(self, config):
-        super().__init__()
-        self.config = config
-        self.num_experts_per_tok = config.num_experts_per_tok
-        # do we use triton kernel for input(activation) quantization or the default dsgemm utils (Pytorch eager based)
-        self.activation_function = MLP.act_fn
-
-        # ep_size is the number of ranks in expert dimension
-        if config.ep_size <= 1:
-            raise ValueError(
-                "For code simplicity, this model only supports distributed experts, "
-                "thus EP size must be > 1, please modify your model config"
-            )
-        self.ep_group = get_group("ep")
-        assert config.ep_size == self.ep_group.size()
-        self.ep_size = config.ep_size
-        self.ep_rank = self.ep_group.rank()
-        self.experts_per_rank = config.n_routed_experts // config.ep_size
-
-        # Use ModuleDict instead of ModuleList to preserve absolute expert
-        # IDs while avoiding `None` experts. The absolute expert IDs match
-        # with checkpoint FQNs.
-        self.experts = nn.ModuleDict()
-        for i in range(self.experts_per_rank):
-            abs_expert_id = self.ep_rank * self.experts_per_rank + i
-            self.experts[str(abs_expert_id)] = MLP(
-                config, intermediate_size=config.moe_intermediate_size
-            )
-        self.gate = MoEGate(config)
-        if config.n_shared_experts is not None:
-            intermediate_size = config.moe_intermediate_size * config.n_shared_experts
-            self.shared_experts = MLP(
-                config=config, intermediate_size=intermediate_size
-            )
-
-        # Group Gemm
-        # Initialize group GEMM strategies if not already loaded
-        if MoE.group_gemm_strategies is None:
-            MoE._initialize_group_gemm_strategies()
-
-        assert (
-            MoE.group_mm in MoE.group_gemm_strategies
-        ), f"selected group gemm {self.group_mm} is not avaiable!"
-        # keep active gg ready
-        self.group_gemm_instance = MoE.group_gemm_strategies[MoE.group_mm]
-        self._buffer_initialized = False
-
-    @classmethod
-    def _initialize_group_gemm_strategies(cls):
-        """Initialize available group GEMM strategies"""
-        cls.group_gemm_strategies = {
-            "torch": TorchBF16GroupGEMM(MLP.act_fn),
-            "torchao": (
-                TorchAOBF16GroupGEMM(MLP.act_fn)
-                if TorchAOBF16GroupGEMM.is_available()
-                else None
-            ),
-            "torchfp8": (
-                TorchFP8GroupGEMM(MLP.act_fn)
-                if TorchFP8GroupGEMM.is_available()
-                else None
-            ),
-            "dsgemm": (
-                DSGroupGEMM(MLP.act_fn, use_triton_quant=True)
-                if DSGroupGEMM.is_available()
-                else None
-            ),
-            "tritoncg": (
-                TritonCGBF16GroupGEMM(
-                    MLP.act_fn,
-                )
-                if TritonCGBF16GroupGEMM.is_available()
-                else None
-            ),
-        }
-
-    def combine_experts(self, submod_name: str):
-        all_weights = []
-        for expert in self.experts.values():
-
-            lin = expert.get_submodule(submod_name)
-            all_weights.append(lin.weight)
-            lin.weight = None
-
-        # let the group gemm strategy prep the final weight layout
-        combined_weight = self.group_gemm_instance.arrange_expert_weights(
-            all_weights, submod_name, self
-        )
-
-        if combined_weight is None:
-            raise NotImplementedError("expert weights not handled by group gemmm")
-
-        self.register_parameter(f"{submod_name}_weight", nn.Parameter(combined_weight))
-
-    # This function is used to create a symm mem buffer for MoE's. It is for
-    # shuffling tokens fully "on-device", as compared to traditional torch
-    # all_to_all APIs which require a GPU-to-CPU sync of the splits.  If a user
-    # calls this function, the `shuffle_method` would switch from
-    # `torch_all_to_all` to `symm_mem`.
-    def setup_symm_mem(self, dtype: torch.dtype, device: torch.device):
-        # Switch shuffle method
-        self.shuffle_method = "symm_mem"
-
-        # Combine expert weights
-        self.combine_experts("gate_proj")
-        self.combine_experts("up_proj")
-        self.combine_experts("down_proj")
-
-        # Assuming worst case, 2x tokens are routed to one EP rank
-        overflow = 2
-        OnDeviceAllToAllV.max_output_len = (
-            self.config.max_seq_len * self.num_experts_per_tok * overflow
-        )
-
-        # Symmetric memory buffers are shared by all MoE instances across
-        # layers, we only need to initialize them once
-        if MoE.token_send_buf is not None:
-            return
-
-        # Input buffer for DP-to-EP shuffle
-        MoE.token_send_buf = symm_mem.empty(
-            self.config.max_seq_len
-            * self.num_experts_per_tok,  # seq len * top k (flattened)
-            self.config.hidden_size,  # hidden dim
-            dtype=dtype,
-            device=device,
-        )
-        # Input buffer for EP-to-DP shuffle
-        MoE.token_gather_buf = symm_mem.empty(
-            self.config.max_seq_len
-            * self.num_experts_per_tok  # seq len * top k (flattened)
-            * overflow,
-            self.config.hidden_size,  # hidden dim
-            dtype=dtype,
-            device=device,
-        )
-
-    def get_send_buf(self):
-        # [Why detach?] During a first forward-backward step, the buffer would
-        # be included in a computational graph. In a second step, autograd will
-        # return an error saying "Trying to backward through the graph a second
-        # time (or directly access saved tensors more than once)". This is
-        # because the buffer is still in the graph, and autograd is trying to
-        # backward through the graph a second time. To avoid this, we detach the
-        # buffer from the graph. `detach()` returns a new tensor, which shares
-        # the same storage with the original one.
-        self.token_send_buf.grad = None
-        return self.token_send_buf.detach()
-
-    def get_gather_buf(self):
-        # See [Why detach?] in `get_send_buf`
-        self.token_gather_buf.grad = None
-        return self.token_gather_buf.detach()
-
-    def forward(self, hidden_states):
-        identity = hidden_states
-        orig_shape = hidden_states.shape
-        # for each token, select top-k experts, and compute the weight for each expert
-        topk_idx, topk_weight = self.gate(hidden_states)
-        hidden_states = hidden_states.view(-1, hidden_states.shape[-1])
-        if self.shuffle_method == "symm_mem":
-            y = self.moe_on_device(hidden_states, topk_idx, topk_weight)
-        else:  # "torch_all_to_all"
-            y = self.moe_forward(hidden_states, topk_idx, topk_weight)
-
-        y = y.view(*orig_shape)
-        if self.config.n_shared_experts is not None:
-            y = y + self.shared_experts(identity)
-        return y
-
-    def moe_forward(self, x, topk_ids, topk_weight):
-        (
-            sorted_tokens,
-            token_indices,
-            tokens_per_expert,
-        ) = self.sort_tokens(x, topk_ids, topk_weight)
-
-        # keep the seqlen dimension for later use without holding onto the sorted tokens
-        seqlen_sorted_tokens = sorted_tokens.shape[0]
-
-        # all to all
-        # This part exchange the information about the number of tokens send and
-        # received by each expert. We can understand this information as "side
-        # band", which is not part of the actual data. Thus no gradient is
-        # needed.
-
-        # Sum the tokens over local experts, then we get tokens per EP rank,
-        # which is the input splits
-        with torch.no_grad():
-            tokens_per_expert_group = tokens_per_expert.new_empty(
-                tokens_per_expert.shape[0]
-            )
-            dist.all_to_all_single(
-                tokens_per_expert_group, tokens_per_expert, group=self.ep_group
-            )
-            input_splits = tokens_per_expert.view(self.ep_size, -1).sum(dim=1)
-
-        # DP to EP token shuffle. This part needs gradient.
-        if self.shuffle_method == "symm_mem":
-            # Move input to the `token_send_buf` symm mem
-            token_send_buf = self.get_send_buf()
-            token_send_buf[: token_indices.shape[0]].copy_(sorted_tokens)
-            # Note: `out=` avoids copy, but it is not differentiable
-            # torch.index_select(x, 0, idxs // topk_ids.shape[1], out=self.token_send_buf[: idxs.shape[0]])
-            token_gather_buf, output_splits = OnDeviceAllToAllV.apply(
-                token_send_buf,
-                input_splits,
-                self.ep_group,
-            )
-            with torch.no_grad():
-                # Received tokens from all other ranks. TODO: use mask instead
-                received = output_splits.sum()
-            # TODO: don't use `received`
-            gathered_tokens = token_gather_buf[:received]
-        else:  # "torch_all_to_all"
-            # Prepare input ans output splits
-            with torch.no_grad():
-                output_splits = tokens_per_expert_group.view(self.ep_size, -1).sum(
-                    dim=1
-                )
-            gathered_tokens = all_to_all_single_autograd(
-                sorted_tokens,
-                output_splits.tolist(),
-                input_splits.tolist(),
-                self.ep_group,
-            )
-
-        # This part prepares a 1D tensor with the same length as
-        # `gathered_tokens`. The 1D tensor is filled with local expert IDs which
-        # the tokens in `gathered_tokens` are headed for. This part doesn't need
-        # gradient.
-        with torch.no_grad():
-            gatherd_idxs = (
-                torch.arange(
-                    tokens_per_expert_group.numel(),
-                    device=tokens_per_expert_group.device,
-                )
-                % self.experts_per_rank
-            )
-            gatherd_idxs = gatherd_idxs.repeat_interleave(tokens_per_expert_group)
-
-        # Prepare buffer for tokens processed by experts
-        if self.shuffle_method == "symm_mem":
-            # Take necessary space from `token_gather_buf` symm mem because we are
-            # going to send them out after expert processing
-            processed_tokens = self.get_gather_buf()[: gathered_tokens.shape[0]]
-        else:  # "torch_all_to_all"
-            processed_tokens = torch.empty_like(gathered_tokens)
-
-        # This part processes the tokens routed to the local experts.
-        # TODO: can we use group GEMM here?
-        for i, expert in enumerate(self.experts.values()):
-            processed_tokens[gatherd_idxs == i] = expert(
-                gathered_tokens[gatherd_idxs == i]
-            )
-
-        # Now shuffle the tokens back to their original owner, i.e. EP to DP shuffle.
-        # The input/output splits are just a reverse of the previous shuffle.
-        if self.shuffle_method == "symm_mem":
-            token_return_buf, _ = OnDeviceAllToAllV.apply(
-                processed_tokens,
-                output_splits,
-                self.ep_group,
-            )
-            returned_tokens = token_return_buf[:seqlen_sorted_tokens]
-        else:  # "torch_all_to_all"
-            returned_tokens = all_to_all_single_autograd(
-                processed_tokens,
-                input_splits.tolist(),
-                output_splits.tolist(),
-                self.ep_group,
-            )
-
-        output_tokens = torch.empty_like(returned_tokens)
-        output_tokens[token_indices] = returned_tokens
-        final_out = (
-            output_tokens.view(*topk_ids.shape, -1)
-            .type(topk_weight.dtype)
-            .mul_(topk_weight.unsqueeze(dim=-1))
-            .sum(dim=1)
-            .type(returned_tokens.dtype)
-        )
-        return final_out
-
-    def sort_tokens(self, x, topk_ids, topk_weights):
-        # This part sorts the token indices so that tokens routed to the same expert reside consecutively.
-        # An implication is that tokens to the same "expert group" (i.e., device) are also consecutive.
-        # Since this is an "aritificial" index creation (final outcome being
-        # `idxs`), we don't need gradients here.
-
-        with torch.no_grad():
-            # [seq_len, n_routed_experts]
-            expert_counts = topk_ids.new_zeros(
-                (topk_ids.shape[0], self.config.n_routed_experts)
-            )
-            # Fill 1 to the selected experts
-            expert_counts.scatter_(1, topk_ids, 1)
-            tokens_per_expert = expert_counts.sum(dim=0)
-            # Token indices for each expert
-            token_indices = topk_ids.view(-1).argsort()
-
-        sorted_tokens = x[token_indices // topk_ids.shape[1]]
-        # assert sorted_tokens.shape == sorted_tokens_shape
-
-        return (sorted_tokens, token_indices, tokens_per_expert)
-
-    # ------- Group GEMM implementation ------
-
-    def _run_group_gemm(self, contig_tokens, m_sizes, m_offsets):
-        """Run the appropriate group GEMM implementation based on configuration"""
-
-        try:
-            return self.group_gemm_strategies[self.group_mm].execute(
-                contig_tokens, m_sizes, m_offsets, self
-            )
-        except Exception as e:
-            # Flag the error
-            print(f"Error using {self.group_mm} strategy: {e}")
-
-    def moe_on_device(self, x, topk_ids, topk_weight):
-        (
-            sorted_tokens,
-            token_indices,
-            tokens_per_expert,
-        ) = self.sort_tokens(x, topk_ids, topk_weight)
-
-        # keep the seqlen dimension for later use without holding onto the sorted tokens
-        seqlen_sorted_tokens = sorted_tokens.shape[0]
-
-        # This part exchange the information about the number of tokens send and
-        # received by each expert. We can understand this information as "side
-        # band", which is not part of the actual data. Thus no gradient is
-        # needed.
-
-        # Sum the tokens over local experts, then we get tokens per EP rank,
-        # which is the input splits
-        with torch.no_grad():
-            tokens_per_expert_group = tokens_per_expert.new_empty(
-                tokens_per_expert.shape[0]
-            )
-            dist.all_to_all_single(
-                tokens_per_expert_group, tokens_per_expert, group=self.ep_group
-            )
-            input_splits = tokens_per_expert.view(self.ep_size, -1).sum(dim=1)
-
-        # Move input to the `token_send_buf` symm mem
-        token_send_buf = self.get_send_buf()
-        token_send_buf[: token_indices.shape[0]].copy_(sorted_tokens)
-        # Note: `out=` avoids copy, but it is not differentiable
-        # torch.index_select(x, 0, idxs // topk_ids.shape[1], out=token_send_buf[: idxs.shape[0]])
-        token_gather_buf, output_splits = OnDeviceAllToAllV.apply(
-            token_send_buf,
-            input_splits,
-            self.ep_group,
-        )
-
-        # We need to permute the received tokens so that tokens for the same expert are contiguous.
-        # This part prepares a 1D tensor `permuted_indices` for such permutation.
-        # This part doesn't need gradient.
-        with torch.no_grad():
-            permuted_indices, m_sizes, m_offsets = generate_permute_indices(
-                tokens_per_expert_group,
-                self.experts_per_rank,
-                self.ep_size,
-                token_gather_buf.shape[0],
-                ALIGN_SIZE_M,
-            )
-
-        # Permute the received tokens so that tokens for the same expert are contiguous.
-        contig_tokens = token_gather_buf[permuted_indices]
-
-        # group gemm - handle all three group gemms (up, gate, down for all experts)
-        hidden_outputs = self._run_group_gemm(
-            contig_tokens,
-            m_sizes,
-            m_offsets,
-        )
-
-        # Prepare buffer for tokens processed by experts
-        processed_tokens = self.get_gather_buf()
-
-        # Move into Symmetric Memory for the return shuffle
-        processed_tokens[permuted_indices] = hidden_outputs
-
-        # Now shuffle the tokens back to their original owner, i.e. EP to DP shuffle.
-        # The input/output splits are just a reverse of the previous shuffle.
-        token_return_buf, _ = OnDeviceAllToAllV.apply(
-            processed_tokens,
-            output_splits,
-            self.ep_group,
-        )
-
-        returned_tokens = token_return_buf[:seqlen_sorted_tokens]
-        output_tokens = torch.empty_like(returned_tokens)
-        output_tokens[token_indices] = returned_tokens
-
-        final_out = (
-            output_tokens.view(*topk_ids.shape, -1)
-            .type(topk_weight.dtype)
-            .mul_(topk_weight.unsqueeze(dim=-1))
-            .sum(dim=1)
-            .type(returned_tokens.dtype)
-        )
-
-        return final_out
-
-
-class Attention(nn.Module):
-    """Multi-headed attention from 'Attention Is All You Need' paper"""
-
-    def __init__(self, config: ModelArgs, layer_idx: Optional[int] = None):
-        super().__init__()
-        self.config = config
-        self.layer_idx = layer_idx
-        self.attention_dropout = config.attention_dropout
-        self.hidden_size = config.hidden_size
-        self.num_heads = config.num_attention_heads
-
-        self.max_position_embeddings = config.max_position_embeddings
-        self.rope_theta = config.rope_theta
-        self.q_lora_rank = config.q_lora_rank
-        self.qk_rope_head_dim = config.qk_rope_head_dim
-        self.kv_lora_rank = config.kv_lora_rank
-        self.v_head_dim = config.v_head_dim
-        self.qk_nope_head_dim = config.qk_nope_head_dim
-        self.q_head_dim = config.qk_nope_head_dim + config.qk_rope_head_dim
-
-        self.is_causal = True
-
-        if self.q_lora_rank is None:
-            self.q_proj = nn.Linear(
-                self.hidden_size, self.num_heads * self.q_head_dim, bias=False
-            )
-        else:
-            self.q_a_proj = nn.Linear(
-                self.hidden_size, config.q_lora_rank, bias=config.attention_bias
-            )
-            self.q_a_layernorm = RMSNorm(config.q_lora_rank)
-            self.q_b_proj = nn.Linear(
-                config.q_lora_rank, self.num_heads * self.q_head_dim, bias=False
-            )
-
-        self.kv_a_proj_with_mqa = nn.Linear(
-            self.hidden_size,
-            config.kv_lora_rank + config.qk_rope_head_dim,
-            bias=config.attention_bias,
-        )
-        self.kv_a_layernorm = RMSNorm(config.kv_lora_rank)
-        self.kv_b_proj = nn.Linear(
-            config.kv_lora_rank,
-            self.num_heads
-            * (self.q_head_dim - self.qk_rope_head_dim + self.v_head_dim),
-            bias=False,
-        )
-
-        self.o_proj = nn.Linear(
-            self.num_heads * self.v_head_dim,
-            self.hidden_size,
-            bias=config.attention_bias,
-        )
-        self._init_rope()
-
-        self.softmax_scale = self.q_head_dim ** (-0.5)
-        if self.config.rope_scaling is not None:
-            mscale_all_dim = self.config.rope_scaling.get("mscale_all_dim", 0)
-            scaling_factor = self.config.rope_scaling["factor"]
-            if mscale_all_dim:
-                mscale = yarn_get_mscale(scaling_factor, mscale_all_dim)
-                self.softmax_scale = self.softmax_scale * mscale * mscale
-
-    def _init_rope(self):
-        if self.config.rope_scaling is None:
-            self.rotary_emb = RotaryEmbedding(
-                self.qk_rope_head_dim,
-                max_position_embeddings=self.max_position_embeddings,
-                base=self.rope_theta,
-            )
-        else:
-            scaling_type = self.config.rope_scaling["type"]
-            scaling_factor = self.config.rope_scaling["factor"]
-            if scaling_type == "linear":
-                self.rotary_emb = LinearScalingRotaryEmbedding(
-                    self.qk_rope_head_dim,
-                    max_position_embeddings=self.max_position_embeddings,
-                    scaling_factor=scaling_factor,
-                    base=self.rope_theta,
-                )
-            elif scaling_type == "dynamic":
-                self.rotary_emb = DynamicNTKScalingRotaryEmbedding(
-                    self.qk_rope_head_dim,
-                    max_position_embeddings=self.max_position_embeddings,
-                    scaling_factor=scaling_factor,
-                    base=self.rope_theta,
-                )
-            elif scaling_type == "yarn":
-                kwargs = {
-                    key: self.config.rope_scaling[key]
-                    for key in [
-                        "original_max_position_embeddings",
-                        "beta_fast",
-                        "beta_slow",
-                        "mscale",
-                        "mscale_all_dim",
-                    ]
-                    if key in self.config.rope_scaling
-                }
-                self.rotary_emb = YarnRotaryEmbedding(
-                    self.qk_rope_head_dim,
-                    max_position_embeddings=self.max_position_embeddings,
-                    scaling_factor=scaling_factor,
-                    base=self.rope_theta,
-                    **kwargs,
-                )
-            else:
-                raise ValueError(f"Unknown RoPE scaling type {scaling_type}")
-
-    def forward(
-        self,
-        hidden_states: torch.Tensor,
-        attention_mask: Optional[torch.Tensor] = None,
-        position_ids: Optional[torch.LongTensor] = None,
-    ) -> Tuple[torch.Tensor, Optional[torch.Tensor], Optional[Tuple[torch.Tensor]]]:
-        bsz, q_len, _ = hidden_states.size()
-
-        if self.q_lora_rank is None:
-            q = self.q_proj(hidden_states)
-        else:
-            q = self.q_b_proj(self.q_a_layernorm(self.q_a_proj(hidden_states)))
-        q = q.view(bsz, q_len, self.num_heads, self.q_head_dim).transpose(1, 2)
-        q_nope, q_pe = torch.split(
-            q, [self.qk_nope_head_dim, self.qk_rope_head_dim], dim=-1
-        )
-
-        compressed_kv = self.kv_a_proj_with_mqa(hidden_states)
-        compressed_kv, k_pe = torch.split(
-            compressed_kv, [self.kv_lora_rank, self.qk_rope_head_dim], dim=-1
-        )
-        k_pe = k_pe.view(bsz, q_len, 1, self.qk_rope_head_dim).transpose(1, 2)
-        kv = (
-            self.kv_b_proj(self.kv_a_layernorm(compressed_kv))
-            .view(bsz, q_len, self.num_heads, self.qk_nope_head_dim + self.v_head_dim)
-            .transpose(1, 2)
-        )
-
-        k_nope, value_states = torch.split(
-            kv, [self.qk_nope_head_dim, self.v_head_dim], dim=-1
-        )
-        kv_seq_len = value_states.shape[-2]
-
-        cos, sin = self.rotary_emb(value_states, seq_len=kv_seq_len)
-
-        q_pe, k_pe = apply_rotary_pos_emb(q_pe, k_pe, cos, sin, position_ids)
-
-        query_states = k_pe.new_empty(bsz, self.num_heads, q_len, self.q_head_dim)
-        query_states[:, :, :, : self.qk_nope_head_dim] = q_nope
-        query_states[:, :, :, self.qk_nope_head_dim :] = q_pe
-
-        key_states = k_pe.new_empty(bsz, self.num_heads, q_len, self.q_head_dim)
-        key_states[:, :, :, : self.qk_nope_head_dim] = k_nope
-        key_states[:, :, :, self.qk_nope_head_dim :] = k_pe
-
-        if attention_mask is not None:
-            # Attention mask was made 4D because the `attn_weights` above is 4D.
-            # We probably can make this mask smarter if we want to pack sequences
-            # together, instead of using padding. This optimization can be used in
-            # inference. For training, if we want to pack sequences, data loader
-            # will pass in a mask containing such info.
-            attention_mask = _prepare_4d_causal_attention_mask(
-                attention_mask,  # None, or user provided mask in 2D
-                (bsz, q_len),
-                hidden_states,
-                0,  # past_key_values_length, 0 when training
-            )
-            if attention_mask.size() != (bsz, 1, q_len, kv_seq_len):
-                raise ValueError(
-                    f"Attention mask should be of size {(bsz, 1, q_len, kv_seq_len)}, but is {attention_mask.size()}"
-                )
-
-        attn_output = torch.nn.functional.scaled_dot_product_attention(
-            query=query_states,
-            key=key_states,
-            value=value_states,
-            attn_mask=attention_mask,
-            dropout_p=self.attention_dropout,
-            is_causal=attention_mask is None,
-            scale=self.softmax_scale,
-        )
-
-        attn_output = attn_output.transpose(1, 2).contiguous()
-        attn_output = attn_output.reshape(bsz, q_len, self.num_heads * self.v_head_dim)
-        attn_output = self.o_proj(attn_output)
-
-        return attn_output
-
-
-class DecoderLayer(nn.Module):
-    def __init__(self, config: ModelArgs, layer_idx: int):
-        super().__init__()
-        self.hidden_size = config.hidden_size
-
-        self.self_attn = Attention(config=config, layer_idx=layer_idx)
-
-        self.mlp = (
-            MoE(config)
-            if (
-                config.n_routed_experts is not None
-                and layer_idx >= config.first_k_dense_replace
-                and layer_idx % config.moe_layer_freq == 0
-            )
-            else MLP(config)
-        )
-        self.input_layernorm = RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
-        self.post_attention_layernorm = RMSNorm(
-            config.hidden_size, eps=config.rms_norm_eps
-        )
-
-    def forward(
-        self,
-        hidden_states: torch.Tensor,
-        attention_mask: Optional[torch.Tensor] = None,
-        position_ids: Optional[torch.LongTensor] = None,
-    ) -> torch.Tensor:
-        """
-        Args:
-            hidden_states (`torch.FloatTensor`): input to the layer of shape `(batch, seq_len, embed_dim)`
-            attention_mask (`torch.FloatTensor`, *optional*):
-                attention mask of size `(batch_size, sequence_length)` if flash attention is used or `(batch_size, 1,
-                query_sequence_length, key_sequence_length)` if default attention is used.
-        """
-        residual = hidden_states
-
-        hidden_states = self.input_layernorm(hidden_states)
-
-        # Self Attention
-        hidden_states = self.self_attn(
-            hidden_states=hidden_states,
-            attention_mask=attention_mask,
-            position_ids=position_ids,
-        )
-        hidden_states = residual + hidden_states
-
-        # Fully Connected
-        residual = hidden_states
-        hidden_states = self.post_attention_layernorm(hidden_states)
-        hidden_states = self.mlp(hidden_states)
-        hidden_states = residual + hidden_states
-
-        return hidden_states
-
-
-Deepseek_INPUTS_DOCSTRING = r"""
-    Args:
-        input_ids (`torch.LongTensor` of shape `(batch_size, sequence_length)`):
-            Indices of input sequence tokens in the vocabulary. Padding will be ignored by default should you provide
-            it.
-
-            Indices can be obtained using [`AutoTokenizer`]. See [`PreTrainedTokenizer.encode`] and
-            [`PreTrainedTokenizer.__call__`] for details.
-
-            [What are input IDs?](../glossary#input-ids)
-        attention_mask (`torch.Tensor` of shape `(batch_size, sequence_length)`, *optional*):
-            Mask to avoid performing attention on padding token indices. Mask values selected in `[0, 1]`:
-
-            - 1 for tokens that are **not masked**,
-            - 0 for tokens that are **masked**.
-
-            [What are attention masks?](../glossary#attention-mask)
-
-            Indices can be obtained using [`AutoTokenizer`]. See [`PreTrainedTokenizer.encode`] and
-            [`PreTrainedTokenizer.__call__`] for details.
-
-            If `past_key_values` is used, optionally only the last `input_ids` have to be input (see
-            `past_key_values`).
-
-            If you want to change padding behavior, you should read [`modeling_opt._prepare_decoder_attention_mask`]
-            and modify to your needs. See diagram 1 in [the paper](https://arxiv.org/abs/1910.13461) for more
-            information on the default strategy.
-
-            - 1 indicates the head is **not masked**,
-            - 0 indicates the head is **masked**.
-        position_ids (`torch.LongTensor` of shape `(batch_size, sequence_length)`, *optional*):
-            Indices of positions of each input sequence tokens in the position embeddings. Selected in the range `[0,
-            config.n_positions - 1]`.
-
-            [What are position IDs?](../glossary#position-ids)
-        past_key_values (`Cache` or `tuple(tuple(torch.FloatTensor))`, *optional*):
-            Pre-computed hidden-states (key and values in the self-attention blocks and in the cross-attention
-            blocks) that can be used to speed up sequential decoding. This typically consists in the `past_key_values`
-            returned by the model at a previous stage of decoding, when `use_cache=True` or `config.use_cache=True`.
-
-            Two formats are allowed:
-            - a [`~cache_utils.Cache`] instance;
-            - Tuple of `tuple(torch.FloatTensor)` of length `config.n_layers`, with each tuple having 2 tensors of
-            shape `(batch_size, num_heads, sequence_length, embed_size_per_head)`). This is also known as the legacy
-            cache format.
-
-            The model will output the same cache format that is fed as input. If no `past_key_values` are passed, the
-            legacy cache format will be returned.
-
-            If `past_key_values` are used, the user can optionally input only the last `input_ids` (those that don't
-            have their past key value states given to this model) of shape `(batch_size, 1)` instead of all `input_ids`
-            of shape `(batch_size, sequence_length)`.
-        inputs_embeds (`torch.FloatTensor` of shape `(batch_size, sequence_length, hidden_size)`, *optional*):
-            Optionally, instead of passing `input_ids` you can choose to directly pass an embedded representation. This
-            is useful if you want more control over how to convert `input_ids` indices into associated vectors than the
-            model's internal embedding lookup matrix.
-        use_cache (`bool`, *optional*):
-            If set to `True`, `past_key_values` key value states are returned and can be used to speed up decoding (see
-            `past_key_values`).
-        output_attentions (`bool`, *optional*):
-            Whether or not to return the attentions tensors of all attention layers. See `attentions` under returned
-            tensors for more detail.
-        output_hidden_states (`bool`, *optional*):
-            Whether or not to return the hidden states of all layers. See `hidden_states` under returned tensors for
-            more detail.
-        return_dict (`bool`, *optional*):
-            Whether or not to return a [`~utils.ModelOutput`] instead of a plain tuple.
-"""
-
-
-class DeepseekModel(torch.nn.Module):
-    """
-    Transformer decoder consisting of *config.num_hidden_layers* layers. Each layer is a [`DecoderLayer`]
-
-    Args:
-        config: ModelArgs
-    """
-
-    def __init__(self, config: ModelArgs):
-        super().__init__()
-        self.config = config
-        self.padding_idx = config.pad_token_id
-        self.vocab_size = config.vocab_size
-
-        # Creating model parts related to my stage
-        assert (
-            config.stage_idx < config.num_stages
-        ), f"Stage {config.stage_idx} is not in the model"
-        print(f"Creating model stage {config.stage_idx} of {config.num_stages}")
-
-        self.embed_tokens = (
-            nn.Embedding(config.vocab_size, config.hidden_size, self.padding_idx)
-            if config.stage_idx == 0
-            else None
-        )
-
-        self.layers = torch.nn.ModuleDict()
-        division = config.num_hidden_layers // config.num_stages
-        residual = config.num_hidden_layers % config.num_stages
-        # Some earlier stages may have 1 more layer than latter stages because
-        # the division may have residual; this is more even than giving the
-        # entire residual to the last stage.
-        layers_per_stage = [
-            division + 1 if stage < residual else division
-            for stage in range(config.num_stages)
-        ]
-        assert sum(layers_per_stage) == config.num_hidden_layers
-        layer_id_start = sum(layers_per_stage[: config.stage_idx])
-        layer_id_end = layer_id_start + layers_per_stage[config.stage_idx]
-        for layer_id in range(layer_id_start, layer_id_end):
-            self.layers[str(layer_id)] = DecoderLayer(config, layer_id)
-
-        self.norm = (
-            RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
-            if config.stage_idx == config.num_stages - 1
-            else None
-        )
-
-        # Initialize weights and apply final processing
-        self.apply(self._init_weights)
-
-    def _init_weights(self, module):
-        std = self.config.initializer_range
-        if isinstance(module, nn.Linear):
-            module.weight.data.normal_(mean=0.0, std=std)
-            if module.bias is not None:
-                module.bias.data.zero_()
-        elif isinstance(module, nn.Embedding):
-            module.weight.data.normal_(mean=0.0, std=std)
-            if module.padding_idx is not None:
-                module.weight.data[module.padding_idx].zero_()
-
-    def forward(
-        self,
-        tokens: torch.Tensor,
-        attention_mask: Optional[torch.Tensor] = None,
-        position_ids: Optional[torch.LongTensor] = None,
-    ) -> torch.Tensor:
-        # Embedding
-        hidden_states = (
-            self.embed_tokens(tokens) if self.embed_tokens is not None else tokens
-        )
-
-        # decoder layers
-        for decoder_layer in self.layers.values():
-            hidden_states = decoder_layer(
-                hidden_states,
-                attention_mask=attention_mask,
-                position_ids=position_ids,
-            )
-
-        hidden_states = (
-            self.norm(hidden_states) if self.norm is not None else hidden_states
-        )
-        return hidden_states
-
-
-class DeepseekForCausalLM(torch.nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        self.model = DeepseekModel(config)
-        self.lm_head = (
-            nn.Linear(config.hidden_size, config.vocab_size, bias=False)
-            if config.stage_idx == config.num_stages - 1
-            else None
-        )
-
-        # Initialize weights and apply final processing
-        # self.post_init()
-
-    def forward(
-        self,
-        tokens: torch.Tensor,
-        attention_mask: Optional[torch.Tensor] = None,
-        position_ids: Optional[torch.LongTensor] = None,
-    ) -> Tuple:
-        r"""
-        Example:
-
-        ```python
-        >>> from transformers import AutoTokenizer, DeepseekForCausalLM
-
-        >>> model = DeepseekForCausalLM.from_pretrained(PATH_TO_CONVERTED_WEIGHTS)
-        >>> tokenizer = AutoTokenizer.from_pretrained(PATH_TO_CONVERTED_TOKENIZER)
-
-        >>> prompt = "Hey, are you conscious? Can you talk to me?"
-        >>> inputs = tokenizer(prompt, return_tensors="pt")
-
-        >>> # Generate
-        >>> generate_ids = model.generate(inputs.input_ids, max_length=30)
-        >>> tokenizer.batch_decode(generate_ids, skip_special_tokens=True, clean_up_tokenization_spaces=False)[0]
-        "Hey, are you conscious? Can you talk to me?\nI'm not conscious, but I can talk to you."
-        ```"""
-        hidden_states = self.model(
-            tokens,
-            attention_mask=attention_mask,
-            position_ids=position_ids,
-        )
-
-        logits = (
-            self.lm_head(hidden_states) if self.lm_head is not None else hidden_states
-        )
-        return logits
-
-    def prepare_inputs_for_generation(
-        self,
-        input_ids,
-        past_key_values=None,
-        attention_mask=None,
-        **kwargs,
-    ):
-        if past_key_values is not None:
-            # Assuming isinstance(past_key_values, Cache):
-            cache_length = past_key_values.get_seq_length()
-            past_length = past_key_values.seen_tokens
-            max_cache_length = past_key_values.get_max_length()
-
-            # Keep only the unprocessed tokens:
-            # 1 - If the length of the attention_mask exceeds the length of input_ids, then we are in a setting where
-            # some of the inputs are exclusivelly passed as part of the cache (e.g. when passing input_embeds as
-            # input)
-            if (
-                attention_mask is not None
-                and attention_mask.shape[1] > input_ids.shape[1]
-            ):
-                input_ids = input_ids[:, -(attention_mask.shape[1] - past_length) :]
-            # 2 - If the past_length is smaller than input_ids', then input_ids holds all input tokens. We can discard
-            # input_ids based on the past_length.
-            elif past_length < input_ids.shape[1]:
-                input_ids = input_ids[:, past_length:]
-            # 3 - Otherwise (past_length >= input_ids.shape[1]), let's assume input_ids only has unprocessed tokens.
-
-            # If we are about to go beyond the maximum cache length, we need to crop the input attention mask.
-            if (
-                max_cache_length is not None
-                and attention_mask is not None
-                and cache_length + input_ids.shape[1] > max_cache_length
-            ):
-                attention_mask = attention_mask[:, -max_cache_length:]
-
-        position_ids = kwargs.get("position_ids", None)
-        if attention_mask is not None and position_ids is None:
-            # create position_ids on the fly for batch generation
-            position_ids = attention_mask.long().cumsum(-1) - 1
-            position_ids.masked_fill_(attention_mask == 0, 1)
-            if past_key_values:
-                position_ids = position_ids[:, -input_ids.shape[1] :]
-
-        model_inputs = {"input_ids": input_ids}
-
-        model_inputs.update(
-            {
-                "position_ids": position_ids,
-                "past_key_values": past_key_values,
-                "use_cache": kwargs.get("use_cache"),
-                "attention_mask": attention_mask,
-            }
-        )
-        return model_inputs
-
-    @staticmethod
-    def _reorder_cache(past_key_values, beam_idx):
-        reordered_past = ()
-        for layer_past in past_key_values:
-            reordered_past += (
-                tuple(
-                    past_state.index_select(0, beam_idx.to(past_state.device))
-                    for past_state in layer_past
-                ),
-            )
-        return reordered_past
-
-    # Setup Symmetric Memory for MoE token shuffle.
-    # Supports inference currently.
-    def setup_symm_mem(self, dtype: torch.dtype, device: torch.device):
-        for layer in self.model.layers.values():
-            if not isinstance(layer.mlp, MoE):
-                continue
-            layer.mlp.setup_symm_mem(dtype, device)
+# For backward compatibility
+model = DeepseekModel
+model_for_causal_lm = DeepseekForCausalLM
